Recovery Lifecycle 2h Step 1: Extract auto-fix worker engine to shared package
