+++ packages/execution-engine/src/index.ts
@@ -32,3 +32,8 @@ export * from './session-driver.js';
 export * from './codex-session-driver.js';
 export * from './claude-session-driver.js';
 export * from './omp-session-driver.js';
+export * from './worker-runtime.js';
+export * from './auto-fix-recovery.js';
+export * from './auto-fix-gating.js';
+export * from './auto-fix-intents.js';
+export * from './lifecycle-events.js';
