+++ packages/execution-engine/tsconfig.json
@@ -17,6 +17,9 @@
     },
     {
       "path": "../data-store"
+    },
+    {
+      "path": "../transport"
     }
   ]
 }
